refactored: Cart Page
