+++ src/pages/Cart.js
@@ -1,48 +1,14 @@
+import { Link } from "react-router-dom";
 import { useCart } from "../dataProvider/CartProvider";
+import { CartCard } from "../Components/CartCard";
 export function Cart() {
-  const { itemsInCart, dispatch: cartDispatch } = useCart();
-  function totalamount(){
-    return itemsInCart.reduce(((acc,value) => value.price * value.quantity + acc),0)
-  }
+  const { itemsInCart } = useCart();
   return (
     <>
       <div className="bg-img cart-bgImg">
         <h2 className="page-heading">Cart</h2>
       </div>
-      <h2 style={{textAlign:"center",margin:"1rem"}}>Total: {totalamount()}</h2>
-      <ul className="card-div">
-        {itemsInCart.map((obj) => {
-          return (
-            <div className="card">
-              <img src={obj.image} alt={obj.name} className="product-img" />
-              <li style={{ margin: "0.5rem" }}>{obj.name}</li>
-              <p style={{ margin: "0.3rem" }}>{obj.price}</p>
-              <div>
-                <button
-                  className="primary-btn"
-                  onClick={() => cartDispatch({ type: "DECREASE_QUANTITY",payload: obj })}
-                  disabled={obj.quantity===1}
-                >
-                  -
-                </button>
-                {obj.quantity}
-                <button
-                  onClick={() => cartDispatch({ type: "INCREASE_QUANTITY",payload: obj })}
-                  className="primary-btn"
-                >
-                  +
-                </button>
-              </div>
-              <button
-                onClick={() => cartDispatch({ type: "REMOVE_FROM_CART",payload: obj })}
-                className="secondary-btn"
-              >
-                Remove
-              </button>
-            </div>
-          );
-        })}
-      </ul>
+      {itemsInCart.length !== 0 ? <CartCard /> : <div style={{margin: "2rem"}}><h6 style={{marginBottom:"1rem",fontSize:"1rem"}}>Your cart is currently empty.</h6> <Link to="/products">Continue browsing here.</Link></div>}
     </>
   );
 }
